+++ github-runner-manager/tests/unit/platform/test_jobmanager_provider.py
@@ -3,19 +3,19 @@
 
 
 """Test for the jobmanager provider module."""
-
+import random
+import secrets
 from typing import Any
 from unittest.mock import MagicMock
 
 import pytest
-from jobmanager_client.models.job import Job
-from jobmanager_client.models.v1_jobs_job_id_health_get200_response import (
-    V1JobsJobIdHealthGet200Response,
+from jobmanager_client import (
+    RunnerHealthResponse,
+    RunnerRegisterResponse,
 )
-from jobmanager_client.models.v1_jobs_job_id_token_post200_response import (
-    V1JobsJobIdTokenPost200Response,
-)
+from jobmanager_client.models.job_read import JobRead
 from jobmanager_client.rest import ApiException, NotFoundException
+from pydantic import BaseModel, HttpUrl
 
 from github_runner_manager.manager.models import InstanceID, RunnerIdentity, RunnerMetadata
 from github_runner_manager.platform.jobmanager_provider import JobManagerPlatform, JobStatus
@@ -26,23 +26,25 @@ from github_runner_manager.platform.platform_provider import (
 )
 from github_runner_manager.types_.github import GitHubRunnerStatus, SelfHostedRunnerLabel
 
+TEST_JOB_MANAGER_URL = "http://jobmanager.example.com"
+
 
 def test_get_runner_context_succeeds(monkeypatch: pytest.MonkeyPatch):
     """
     arrange: Mock the client api to return a token.
     act: Call get_runner_context.
-    assert: The correct token and the correct runner are returned..
+    assert: The correct token and the correct runner are returned.
     """
     call_api_mock = MagicMock()
     monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
-    call_api_mock.return_value = V1JobsJobIdTokenPost200Response(token="mytoken")
+    runner_id = random.randint(1, 1000)
+    token = secrets.token_hex(16)
+    call_api_mock.return_value = RunnerRegisterResponse(id=runner_id, token=token)
 
-    metadata = RunnerMetadata(
-        platform_name="jobmanager", runner_id="3", url="http://jobmanager.example.com"
-    )
+    metadata = RunnerMetadata(platform_name="jobmanager", runner_id=None, url=None)
     instance_id = InstanceID.build(prefix="unit-0")
     labels = ["label"]
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(url=TEST_JOB_MANAGER_URL)
 
     context, runner = platform.get_runner_context(metadata, instance_id, labels)
 
@@ -56,7 +58,11 @@ def test_get_runner_context_succeeds(monkeypatch: pytest.MonkeyPatch):
 @pytest.mark.parametrize(
     "api_return_value, error_message",
     [
-        pytest.param(V1JobsJobIdTokenPost200Response(), "Empty token", id="No token"),
+        pytest.param(
+            RunnerRegisterResponse(id=random.randint(1, 10), token=""),
+            "Empty token",
+            id="Empty token",
+        ),
         pytest.param(ApiException, "API error", id="Exception from api"),
     ],
 )
@@ -77,7 +83,7 @@ def test_get_runner_context_fails(
     )
     instance_id = InstanceID.build(prefix="unit-0")
     labels = ["label"]
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(url=TEST_JOB_MANAGER_URL)
 
     with pytest.raises(PlatformApiError) as exc:
         _context, _runner = platform.get_runner_context(metadata, instance_id, labels)
@@ -88,8 +94,28 @@ def test_get_runner_context_fails(
 @pytest.mark.parametrize(
     "api_return_value, picked_up",
     [
-        pytest.param(Job(status=JobStatus.IN_PROGRESS.value), True, id="in progress job"),
-        pytest.param(Job(status=JobStatus.PENDING.value), False, id="pending job"),
+        pytest.param(
+            JobRead(
+                status=JobStatus.IN_PROGRESS.value,
+                architecture="arm64",
+                base_series="jammy",
+                id=1,
+                requested_by="foobar",
+            ),
+            True,
+            id="in progress job",
+        ),
+        pytest.param(
+            JobRead(
+                status=JobStatus.PENDING.value,
+                architecture="arm64",
+                base_series="jammy",
+                id=1,
+                requested_by="foobar",
+            ),
+            False,
+            id="pending job",
+        ),
     ],
 )
 def test_check_job_been_picked_up(monkeypatch: pytest.MonkeyPatch, api_return_value, picked_up):
@@ -102,11 +128,23 @@ def test_check_job_been_picked_up(monkeypatch: pytest.MonkeyPatch, api_return_va
     monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
     call_api_mock.side_effect = [api_return_value]
 
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(TEST_JOB_MANAGER_URL)
     metadata = RunnerMetadata(
         platform_name="jobmanager", runner_id="3", url="http://jobmanager.example.com"
     )
-    job_url = "http://example.com/v1/jobs/3"
+
+    # we use a BaseModel to convert a url string to a HttpUrl
+    class JobUrlModel(BaseModel):
+        """Model for job URL.
+
+        Attributes:
+            url: The URL of the job.
+        """
+
+        url: HttpUrl
+
+    # we can pass a string here, mypy doesn't understand it
+    job_url = JobUrlModel(url="http://jobmanager.com/v1/jobs/1234").url  # type: ignore
 
     assert platform.check_job_been_picked_up(metadata, job_url) == picked_up
 
@@ -121,14 +159,93 @@ def test_check_job_been_picked_fails(monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
     call_api_mock.side_effect = ApiException
 
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(TEST_JOB_MANAGER_URL)
     metadata = RunnerMetadata(
         platform_name="jobmanager", runner_id="3", url="http://jobmanager.example.com"
     )
-    job_url = "http://example.com/v1/jobs/3"
+
+    # we use a BaseModel to convert a url string to a HttpUrl
+    class JobUrlModel(BaseModel):
+        """Model for job URL.
+
+        Attributes:
+            url: The URL of the job.
+        """
+
+        url: HttpUrl
+
+    # we can pass a string here, mypy doesn't understand it
+    job_url = JobUrlModel(url="http://jobmanager.com/v1/jobs/1234").url  # type: ignore
 
     with pytest.raises(PlatformApiError):
         platform.check_job_been_picked_up(metadata, job_url)
+
+
+@pytest.mark.parametrize(
+    "job_url, expected_msg",
+    [
+        pytest.param(
+            "http://jobmanager.com/v1/runner",
+            'Job URL path does not start with "/v1/jobs/"',
+            id="wrong path",
+        ),
+        pytest.param(
+            "http://jobmanager.com/",
+            'Job URL path does not start with "/v1/jobs/"',
+            id="no path",
+        ),
+        pytest.param(
+            "http://jobmanager.com",
+            'Job URL path does not start with "/v1/jobs/"',
+            id="no path and no trailing slash",
+        ),
+        pytest.param(
+            "http://jobmanager.com/v1/jobs/",
+            "Job URL path does not contain a valid job_id after '/v1/jobs/'",
+            id="job id missing",
+        ),
+        pytest.param(
+            "http://jobmanager.com/v1/jobs/",
+            "Job URL path does not contain a valid job_id after '/v1/jobs/'",
+            id="job id non-int",
+        ),
+    ],
+)
+def test_check_job_been_picked_up_job_url_validation_err(
+    job_url: str, expected_msg: str, monkeypatch: pytest.MonkeyPatch
+):
+    call_api_mock = MagicMock()
+    monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
+    call_api_mock.side_effect = [
+        JobRead(
+            status=JobStatus.IN_PROGRESS.value,
+            architecture="arm64",
+            base_series="jammy",
+            id=1,
+            requested_by="foobar",
+        )
+    ]
+
+    platform = JobManagerPlatform(TEST_JOB_MANAGER_URL)
+    metadata = RunnerMetadata(platform_name="jobmanager", runner_id="3", url=job_url)
+
+    # we use a BaseModel to convert a url string to a HttpUrl
+    class JobUrlModel(BaseModel):
+        """Model for job URL.
+
+        Attributes:
+            url: The URL of the job.
+        """
+
+        url: HttpUrl
+
+    # we can pass a string here, mypy doesn't understand it
+    job_url = JobUrlModel(url=job_url).url  # type: ignore
+
+    with pytest.raises(ValueError) as exc_info:
+        platform.check_job_been_picked_up(metadata, job_url)
+
+    assert expected_msg in str(exc_info.value)
 
 
 @pytest.mark.parametrize(
@@ -156,14 +273,17 @@ def test_get_runner_health(
     call_api_mock = MagicMock()
     monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
 
-    api_return_value = V1JobsJobIdHealthGet200Response(
+    api_return_value = RunnerHealthResponse(
         label="label",
         status=job_status,
         deletable=job_deletable,
+        cpu_usage="10%",
+        ram_usage="20%",
+        disk_usage="30%",
     )
     call_api_mock.side_effect = [api_return_value]
 
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(url=TEST_JOB_MANAGER_URL)
     instance_id = InstanceID.build(prefix="unit-0")
     metadata = RunnerMetadata(
         platform_name="jobmanager", runner_id="3", url="http://jobmanager.example.com"
@@ -206,10 +326,13 @@ def test_get_runner_health(
                 ),
             ],
             [
-                V1JobsJobIdHealthGet200Response(
+                RunnerHealthResponse(
                     label="label",
                     status="IN_PROGRESS",
                     deletable=False,
+                    cpu_usage="10%",
+                    ram_usage="20%",
+                    disk_usage="30%",
                 ),
                 ApiException,
             ],
@@ -248,10 +371,13 @@ def test_get_runner_health(
                 ),
             ],
             [
-                V1JobsJobIdHealthGet200Response(
+                RunnerHealthResponse(
                     label="label",
                     status="FAILED",
                     deletable=True,
+                    cpu_usage="10%",
+                    ram_usage="20%",
+                    disk_usage="30%",
                 ),
                 NotFoundException,
             ],
@@ -291,7 +417,7 @@ def test_get_runners_health(
     monkeypatch.setattr("jobmanager_client.ApiClient.call_api", call_api_mock)
     call_api_mock.side_effect = jobmanager_side_effects
 
-    platform = JobManagerPlatform()
+    platform = JobManagerPlatform(TEST_JOB_MANAGER_URL)
     runners_health_response = platform.get_runners_health(requested_runners)
 
     expected_health_response = expected_health_response