+++ docs/changelog.md
@@ -1,10 +1,22 @@
 # Changelog
 
+<<<<<<< minor-dashboard-improvement
 
 ### 2025-06-17
 
 - Fix bug where SSH connection error always appears in the logs.
 - Change the repository panel to show the number of jobs per workflow per repository.
+=======
+This changelog documents user-relevant changes to the GitHub runner charm.
+
+## 2025-06-24
+
+- Fix a bug where deleted GitHub Actions Job would cause an endless loop of retries.
+
+### 2025-06-17
+
+- Fix a bug where SSH connection error always appears in the logs.
+>>>>>>> main
 
 ### 2025-06-16
 
@@ -43,9 +55,9 @@ network/infrastructure specific things.
 
 ### 2025-05-06
 
-- The ssh health checks are removed and the platform providers (GitHub or the JobManager) are used instead to get the runners health
+- The ssh health checks are removed and GitHub is used instead to get the runners health
 information. This implies many changes in both the structure of the project and its functionality. Potentially, many race conditions should
-disappear for the GitHub case. 
+disappear.
 
 ### 2025-04-28